+++ test/models/models.jl
@@ -95,10 +95,6 @@ end
 @testset "Test for errors" begin
     @test_throws AssertionError Models.Model(MLP(); likelihood=:regression)
 
-    # test Flux models
-    @test_throws ArgumentError Models.FluxModel("dummy"; likelihood=:regression)
-    @test_throws ArgumentError Models.FluxEnsemble("dummy"; likelihood=:regression)
-
     data = TaijaData.load_linearly_separable()
     counterfactual_data = CounterfactualExplanations.DataPreprocessing.CounterfactualData(
         data[1], data[2]